Function sui_common::authority_aggregation::quorum_map_then_reduce_with_timeout

pub async fn quorum_map_then_reduce_with_timeout<'a, C, K, Client: 'a, S: 'a, V: 'a, R: 'a, E, FMap, FReduce>(
    committee: Arc<C>,
    authority_clients: Arc<BTreeMap<K, Arc<Client>>>,
    initial_state: S,
    map_each_authority: FMap,
    reduce_result: FReduce,
    initial_timeout: Duration
) -> Result<(R, FuturesUnordered<impl Future<Output = (K, Result<V, E>)> + 'a>), S>where
    K: Ord + ConciseableName<'a> + Clone + 'a,
    C: CommitteeTrait<K>,
    FMap: FnOnce(K, Arc<Client>) -> AsyncResult<'a, V, E> + Clone + 'a,
    FReduce: Fn(S, K, StakeUnit, Result<V, E>) -> BoxFuture<'a, ReduceOutput<R, S>> + 'a,

This function takes an initial state, than executes an asynchronous function (FMap) for each authority, and folds the results as they become available into the state using an async function (FReduce).

FMap can do io, and returns a result V. An error there may not be fatal, and could be consumed by the MReduce function to overall recover from it. This is necessary to ensure byzantine authorities cannot interrupt the logic of this function.

FReduce returns a result to a ReduceOutput. If the result is Err the function shortcuts and the Err is returned. An Ok ReduceOutput result can be used to shortcut and return the resulting state (ReduceOutput::End), continue the folding as new states arrive (ReduceOutput::Continue), or continue with a timeout maximum waiting time (ReduceOutput::ContinueWithTimeout).

This function provides a flexible way to communicate with a quorum of authorities, processing and processing their results into a safe overall result, and also safely allowing operations to continue past the quorum to ensure all authorities are up to date (up to a timeout).